requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
openai==1.78.1
priority==2.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==7.2.9
PyYAML==6.0.2
Quart==0.20.0
quart-cors==0.8.0
rich==14.0.0
sniffio==1.3.1
tls-client==1.0.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
wsproto==1.2.0
